+++ Assets/Dust/CHANGELOG.md
@@ -6,6 +6,7 @@
 - DuClampFactoryMachine: added flags to turn on/off PSR clamp for each or XYZ-axises 
 - On add Du*Event component it correct append to Undo stack
 - DuFactoryInstance: allow add only one component to object (Issue #24)
+- DuFactory: forced rebuild factory instances on change any param of depended prefabs (Issue #18)
 
 
 ________________________________________________________________________________________________________________________